Moves FormFieldDropdownSingleSelection
