+++ src/components/Field/FieldTypes.tsx
@@ -4,7 +4,7 @@ import { FieldDefCheckbox } from "@root/components/Field/FormFieldCheckbox";
 import { FieldDefChip } from "@root/components/Field/FormFieldChipSingleSelect";
 import { FieldDefColor } from "@root/components/Field/FormFieldColorPicker/ColorPickerTypes";
 import { FieldDefDate } from "@root/components/Field/FormFieldDate/DateField";
-import { FieldDefDropdown } from "@root/forms/FormFieldDropdownSingleSelection";
+import { FieldDefDropdown } from "@root/components/Field/FormFieldDropdownSingleSelection";
 import { FieldDefImageUpload } from "@root/forms/FormFieldImageUpload";
 import { FieldDefImageVideoDocumentLink } from "@root/forms/FormFieldImageVideoLinkDocumentBrowsing";
 import { FieldDefMapCoordinates } from "@root/forms/FormFieldMapCoordinates";
